requirements.txt
pyyaml
requests
bs4
lxml
requests_cache
tabulate
pyfiglet
markdown 